+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Fix the modal detection
     - Let the caller choose the classes to apply to the modal
     - Let the caller change the configuration of existing tabs
+    - Have a messager as soon as we run in a modal
 
 ### 1.2.0